Protocol buffer messages associated with the next architecture
This change includes a new fabric.proto file that defines
messages between the SDK, endorsers, ordering service, and
committers. These messages are required by the next architecture.
The messages have been derived from the Hyperledger Protocol
Working Group and many others working on the next architecture.
This API is not the final form, but a proposed start that
can be used and evolved as the new components are built.

Appended "2" to those structs that are common with fabric.proto.
This allows us fabric-next to co-exist with fabric and helps
keep track of structures that would need to be renamed as
integration progresses.

Also introduced a ID field in Proposal to help make transition
from current "Transaction" processing code in the fabric smoother
(such as in chaincode execution paths which relies on a txid to
keep track of transactions currently being processed).

Added fabric-next_test.go to test for exact match for fields
Change-Id: I42130175d6a81dc5de3d92c6adba7de0d3266bed
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>