Prevent Jenkins build from cleaning baseimage
Jenkins build runs 'make dist-clean' before it begins the build
process. This has the unfortunate side effect of deleting the
fabric-baseimage docker image in the local repository. Hence,
while we enhanced the make targets so that we could leverage
a baseimage previously built, it is not benefiting CI.

Added a 'scrub' target that does clean the baseimage, and
removed 'base' from the set of image targets cleaned, normally.

Fixes FAB-255

Change-Id: Ib9087211e30e2cf74c8a69caf26370c9fc57e60d
Signed-off-by: Christopher Ferris <chrisfer@us.ibm.com>