Add a fileledger implementation in rawledger.
This addition supplements the existing ramledger rawledger
implementation.  Neither this, nor the ramledger is intended to be
performant, but they are both intended to be extremely simple.

In the case of this added fileledger, it is intended to allow data
persistence between stopping and starting of an orderer.

Ultimately, the expectation would be that for the performant case,
a high performance database of some sort, such as RocksDB or other
would likely be used for persisting data.

https://jira.hyperledger.org/browse/FAB-326

Change-Id: Ic75bee7cd27b311b512b2c3e0e9741b40baffbf1
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>