chaincode life-cycle system chaincode for a chain
The life-cycle system chaincode (lccc) manages chaincodes for a chain in an
endorser. The life-cycle is basically the "deploy", "upgrade", "stop"
and "start" actions. This changeset provides the basic chaincode for
creating the table of chaincodes and implements just the "deploy" command.
This work will be developed till the basic endorser functions are fully
implemented.

This driver for this chaincode will be checked in the next changeset.

NOTE - this change also fixes the limitation where only one system chaincode
can be running at a time.

This is part of the feature development of FAB-181, FAB-182, FAB-183.

Change-Id: Iff36fee7c5b9a9ce4658910db73304a6bcd7e3d4
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>