Add orderer config mechanism
This changeset adds an orderer.yaml config file which
specifies the defaults which were previously hardcoded.

It also correspondingly removes the hardcoded defaults and
utilizes the configuration in those places.

It resolves https://jira.hyperledger.org/browse/FAB-386

This configuration is being done using a variant of Viper's
Unmarshal.  In newer versions of Viper there is a new
function UnmarshalExact which throws errors when Unmarshaling
encounters config fields which are unexpected.

However, there are two outstanding bugs around this Viper
feature which affect us and do not make pulling the newer code
worthwhile at this point.

1. Unmarshaling does not appropriately support time durations

https://github.com/spf13/viper/issues/105
https://github.com/spf13/viper/pull/205

2. Unmarshaling does not correctly handle env overrides of
nested config parameters

https://github.com/spf13/viper/issues/160

This changeset includes a stand-in implementation of UnmarshalExact
which does not suffer from these Viper bugs.  These workarounds
should be removed once fixes are pushed upstream.

Committing to reclaim this commit after a little Gerrit accident

Change-Id: I931b955b0d8fdaacb240a1b480eb695109774e38
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>