Urgent logging fix for gossip
This is an urgent logging fix to a bug that might break gossip tests.
in op/go-logging/backend.go,
the log invokes IsEnabledFor which goes to a global field defaultBackend
that IsEnabledFor method for that global field invokes GetLevel
which reads from a map.

If a logger is initalized at the same time another goroutine reads from the map,
a concurrent read-write would occur and will result in panic.

Change-Id: Ief788f3da722aa38b7c743179d6dd1a1dc60936a
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>