Gossip Discovery small fix
Fixes a scenario in which a node is resurrected, but isn't found in the dead list.
Then it needs to be added to the alive list regardless.

Also fixes a bug in which when expiring a peer,
if it has no recent seen timestamp- do not move it
to the lastDeadTS because no such exists.

Change-Id: I7b5fdd21abc4dffcf2bcc03c0d9a9190d19d3506
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>