FAB-437 bare-minimum, end to end skeleton using solo
The main purpose of this checking
  . show commit followed by simulation in action
  . users can get a feel for the flow of the protocol across
    the different legs of the end-end path
  . identify key areas for attention (esp. grep for
    "!!IMPORTANT!!")

"deploy and "invoke" Chaincode commands from CLI will also
convert a successful proposal response into a transaction
and send it to the Orderer (if configured in core.yaml).

"query" is removed from CLI. Invoke can also return values
now in ProposalResponse.Response.Payload.

REST calls should not be affected and should work with
old ledger.

See core.yaml for default orderer setup.

This also introduces a stop-gap "committer" whose only
task is to commit blocks from the orderer.

To test :
1. Terminal 1 - run the "solo" orderer
   cd fabric/orderer
   go build
   ./orderer

2. Terminal 2 - run the peer
   peer node start 1>/tmp/peer.out 2>&1

3. Terminal 3 - deploy and invoke take usual params
   peer chaincode deploy ... 1>/tmp/out 2>&1
   peer chaincode invoke ... 1>/tmp/out 2>&1

/tmp/peer.out and /tmp/out will contain tell tale signs
of the round trip in action.

Change-Id: Ic1aa31993fc57ce145c39967d4d682fd2dc5704b
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>