Orderer broadcast filtering framework
In order to allow reconfiguration messages to be supported by the
orderer as well as to allow authentication of message signatures from
clients, the ordering service will need to support limited filtering of
broadcast messages as they arrive.

This changeset defines a simple interface for doing so as well as a
structure which supports applying many rules at once.  The solo orderer
is hooked into this framework, but for the moment only supports the
existing behavior of 'Accept all non-empty messages', this will change
once other more sophisticated filters are added.

The hope is that this filtering framework can be re-used across ordering
plugins, especially the component for verifying client signatures on
incoming messages.

https://jira.hyperledger.org/browse/FAB-591

Change-Id: I534330a1cb2b675c6f34e73079654ab3280520c0
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>