Add trivial signature validation dsl via protobuf
This changeset attempts to make a first pass at having a generic enough
to be useful, but not so generic as to be impossible to understand
domain specific language (via protobuf) to express cryptographic
validation schemes.

In particular, the two primitives which comprise a policy are:
    NOutOf(n, []policies)
    SignedBy(id)

Please note that this DSL is relying on the structure imposed by
protobuf, and therefore defines the entire grammar of the DSL in 10
lines.  There is an additional envelope message to allow the
specification to be versioned.

This was developed especially for aiding in specifying bootstrap
configuration for signature policies, however, its applicability for
other areas such as endorsement seems likely.

https://jira.hyperledger.org/browse/FAB-704

Change-Id: I330b0660caf90b09034e5a1c167c08a5c2078e8f
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>