Nominate Jason Yellick as fabric maintainer
I'm nominating Jason to become a fabric maintainer. His work is
consistently of the highest quality. His commits are always
well documented, and complete with tests and in-line documentation
as appropriate.

His reviews are always high quality constructive criticism.
I think he'll make a great addition to the team and provide another
expert in consensus domain.

Change-Id: If6d0628e5170f51268ff55d81d9d8f0540d24d86
Signed-off-by: Christopher Ferris <chrisfer@us.ibm.com>