Add a Chain Configuration Manager
In order to specify genesis configuration, it must be encoded in the
block in a consumable way.  This configuration must then be stored
in a way accessible to any other components which wish to consume this
configuration.  Further, when new configuration is proposed it must be
validated, and applied or rejected depending on its validity.

This changeset creates a configuration manager which handles all these
tasks.  It allows for registering arbitrary configuration handlers for
different configuration types, and has been designed in concert with the
policy manager to allow the policy manager to be plugged in as a
configuration handler so that policy both enforces what configuration
modifications are allowed and is itself a type of configuration.

Because of the complicated nature of verifying that a configuration is
valid (well formed, not forged or replayed, conforms to existing policy,
and satisfies the assorted handlers) this changeset is unfortunately
large, but is needed to achieve complete code and test coverage of
invalid configurations.

This resolves:

https://jira.hyperledger.org/browse/FAB-724

Change-Id: I44cbc3f2e5850d8c22977c923c044d353a9bc453
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>