Revert "... uses hardcoded hashcode for example02"
This reverts commit 5d9a3eabfd2030a38966a7f02e032b6ce7492d4a.

chaincode_example04 calls chaincode_example02 via a hard-coded ID.
This is intentional to catch change to the ID when NOTHING in
chaincode_example02 or in its folder changes.  ie, a test for the
constancy of the hash generation algorithm.

The patch attempt pass the ID as a parameter to prevent accidental
changes to chaincode_example02 from causing test failures. While
well-intented, this subverts the above reasoning behind the
hard-coding.

Also note that chaincode_example05 does take the argument as a
parameter (so that usage is already covered).

Change-Id: I8e305efb6fd577bc7d7e3d720d293cfb0e3c191f
Signed-off-by: Greg Haskins <gregory.haskins@gmail.com>