Gossip internal API changes, and discovery refact
This commit prepares the gossip component to disregard a peer's ID,
and to use only PKI-ID as identification instead,
since it's the only thing that can be tied to its identity.

This commit also contains some additional protobuf messages
needed for further commits, and also has a partial refactoring
of the discovery module to accomodate the internal API changes.

Change-Id: I6c12bee5fbb072bba0ab34fb4a09f5025b106161
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>