FAB-186 Implementation of Endorser and ESCC logic
https://jira.hyperledger.org/browse/FAB-186

This push-request implements the endorser and ESCC logic that are exercised
when a proposal is received by the peer. There are still a few missing pieces,
namely signature generation and verification, verification of the validity of
the signing certificate and replay attack protection. They will be added as
soon as COP/BCSSP interfaces are ready.

Change-Id: I35ecd3306605aca6c93399363648d228f397896a
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>