Merge orderer protos into main protos
This commit begins the work of codifying a single set of common protos
which both the orderer and fabric access, plus a specific set of orderer
protos which are necessary to interact with the ordering service.  In
the future, it is expected that the peer protos will go into their own
directory (importing the common protos).

This changeset looks a little scarier than it is, because the effects of
changing these paths are far reaching indeed, but should be relatively
easy to review.

Note that this is the first ever cross-directory-dependent protos
included in the fabric.  This means the generation incantations change a
little.  Instead of importing the directory which contains the protos,
instead, a common root directory must be picked.

The compile_protos.sh has been changed to look for .protoroot files and
if present use this as the proto compilation root, rather than the
directory containing the proto.

Change-Id: I619ea2c50ffd450640dafbddc7d7e5f34f809f3b
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>