Update common/orderer protos
This changeset incorporates feedback on the move towards sharing common
proto definitions between the orderer and the peers.  In particular, it
splits the Header into a ChainHeader and a SignatureHeader, where the
ChainHeader is usually common to a transaction or configuration item
while the SignatureHeader provides the fields necessary to provide a
secure signature (an identity, nonce, and epoch). (But the signature is
usually over both plus some other data)

It also incorporates some other changes, like making the type not be
enumerated (with reserved fields 0-10k) but and moves the existing
enumerations out of the message so that their integer equivalent can
still be shared across implementations.

Change-Id: Ia8fdb253bb42dc25a18c05100642ec942d9253ec
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>