Replace RocksDB by goleveldb
https://jira.hyperledger.org/browse/FAB-788
RocksDB has a patent infringement license in it from Facebook
(https://github.com/facebook/rocksdb/blob/master/PATENTS)
Many users may not be comfortable with the term of the license.

The alternatives include
1. LevelDB (https://github.com/google/leveldb) with a go wrapper (https://github.com/jmhodges/levigo),
2. goleveldb (https://github.com/syndtr/goleveldb) - a porting of leveldb in golang
3. BoltDB (https://github.com/boltdb/bolt)

BoltDB is suitable for read heavy workloads (e.g., LDAP)
but has a relatively poor performance for read-write workloads.
Of the other two options, goleveldb is chosen because it is
implemented in golang and hence easy to intergate and maintain.
In addition, as a precedent, ethereum go implementation also
uses this package
https://github.com/ethereum/go-ethereum/blob/master/ethdb/database.go

Change-Id: Ia4fb5a6f9299e613d03d8b414a51bf479bfafd59
Signed-off-by: manish <manish.sethi@gmail.com>