Move simplebft to orderer and fix proto compile
Proto compilation was broken for sbft.  In order to fix this, the
simplebft.proto needed to be under the same proto root as the
config.proto and consensus.proto from the sbft package.

This seemed like an opportune time to move the simplebft code out of the
legacy consensus dir and into orderer dir.

Change-Id: Ie84dfaede6ebd5a2cffdcd021bf1d2f1966979f2
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>