Remove fields from ConfigurationEnvelope
The ConfigurationEnvelope used to contain a ChainID and a Sequence
number.  These fields were convenient, but ultimately inferrable from
the configuration envelope contents.  In an effort to make this message
easier to understand and assemble, this changeset removes the sequence
and chainID from the ConfigurationEnvelope.

It also correspondingly fixes the configuration transaction processing
code to examine the contents of the ConfigurationEnvelope to infer these
values.

Change-Id: I9336360071c78d163506491bccb8e51c87e8fce6
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>