FAB-1087 Add config option in core.yaml for history
Add config option in core.yaml for history database
to store the transaction history.  The history database
will be a queryable database such as CouchDB.

The default will be false.  If set to false, the transaction
history will not be persisted in the history database.

Move the ledger config directory up one level to ledger
to be the config for state and history databases as well
as any other ledger functionality that needs a config option.

Change-Id: If4607047082b686dd903d8b3805362d0c05e1c64
Signed-off-by: Mari Wade <mariwade@us.ibm.com>