FAB-872 Multichannel support: message extension
https://jira.hyperledger.org/browse/FAB-872
This commit extends the gossip message to introduce 2 fields:
1) Channel - obviously, since we'll do per-channel routing
2) Tag     - This is for disseminating information inside an organization,
             and for specifying per-channel routing is enabled

This commit also refactors the code related to
the gossip internal message store comparison.

Last but not least- since state transfer is going to be per channel,
the commit introduces a new message type - StateInfo that'll be used
for the state transfer mechanism but per-channel and not globally.

Change-Id: Ibba0b5494ef4ef35420e2c6b09c058f17e56db0c
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>