[FAB-798] Factor out block cutting logic
This changeset removes the block cutting logic from solo and pushes it
into a common package to be reused by other components.

It is designed specifically to support applications which cannot utilize
a batch timer (Kafka) as well as applications for which cutting a block
does not imply committing it (SBFT).

Change-Id: Iea14eb1c4a030406cfe4a19f691464c2805b23ee
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>