[FAB-1161] Push genesis block upon orderer init
Before this changeset, the genesis block would be pushed with the first
broadcast request. A side-effect of this is that any deliver requests
reaching the orderer before the first broadcast request would panic.

This changeset fixes that by having the orderer push the genesis block
(and create the respective Kafka topic) upon initialization.

Change-Id: Ice5d84b45bcd1565f916f8ac13073380090e6036
Signed-off-by: Kostas Christidis <kostas@christidis.io>