Define a static TestChainID
As multichain support comes online, the randomly generated chainID in
the static bootstrapper causes problems.  This is because the clients
must know the chainID ahead of time before submitting transactions.
This has also caused problems for SBFT testing as multiple processes
must agree on the system chain ID.

This changeset defines a static TestChainID for static bootstrapping and
fixes the assorted sample clients to utilize this TestChainID.

Change-Id: I73457faea4b29fc6f8f62a196f419fcc66cb36ed
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>