Hook multichain manager into main path
The previous changeset created the multichain manager but left it
entirely out of the execution path.  This changeset hooks the multichain
manager into the primary path and removes much of the redundant code it
replaces.

This changeset only supports solo as a consumer of the multichain
manager, but kafka will follow next (as the multichain manager gives
multichain support for free to implementers of its Consenter interface).

Change-Id: I4e5b19971afe5b434843f3e0011fd18df098aa0f
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>