FAB-1198-rm old pb.Transaction, pb.Block
NOTE - Removing of old proto.Transaction is the cause for
the large change set. It affects chaincode framework and
all users of the framework such as endorser, system chaincodes,
unit tests etc.

Transaction2 is renamed to Transaction.
Response2 is renamed to Response.
Message2 is renamed to Message.

The changes are fully described in
    https://jira.hyperledger.org/browse/FAB-1198

Summary
=======
   . Remove old Transaction and rename Transaction2
   . Cleanup of Chaincode protobuf message
   . Add TxID for SDK and higher layers to optionally
     set (currently errors if not set)

ChaincodeMessage removes QUERY and QUERY_CHAINCODE enums.

Shim interface does not enforce Query or QueryChaincode.

chaincode_example02 and 05 implement Query function via
the Invoke implementation.

The "noop" system chaincode is removed
   . it was using Transaction which is not an endorser
     artifact any longer
   . there are many system chaincodes to that thoroughly
     test sys chaincode functions

Change-Id: Ib77b7e5a6756eac47e888309816076580ae505e7
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>