FAB-1045 Gossip pull refactoring: uint64->string
Currently the gossip pull mechanism can only handle items
uint64 as keys (for digests).
I refactored this to be string, because a string can hold
both ints, hashes and can also be used as a key in a map.

This is needed for disseminating entities such as certificates
of peers.

Change-Id: I06be757275dccd4e78f055bb703aeadd52787c33
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>