Move configfilter to configtx package
The broadcastfilter package had a subpackage called configfilter.  This
was wrong for a number of reasons.  First off, the package name
stuttered, and secondly it does not filter configuration, rather, it
filters configuration transactions.  Finally, because of the close
dependency on the configtx package, it is natural to group the two
together.

This move is meant to serve as a prelude to introducing a multichain
broadcastfilter which will live within the multichain package.

Change-Id: Ibfe2ccc4d5285333ef1d745597ae57641c1c425c
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>