fix sbft backlog processing
sbft backlog needs to be processed on every phase transition. Otherwise,
messages put in the backlog may be as good as lost, in case, e.g., a
pre-prepare is received after corresponding prepare/commit/checkpoint msgs.

This is demonstrated by the new test TestMsgReordering (which would so far
fail) in simplebft_test.go.

Change-Id: I4b38a922e5d565c8e7a92a66f3e125fd30e8715f
Signed-off-by: Marko Vukolic <mvu@zurich.ibm.com>