[FAB-1279] Add dynamic chain creation path
This changeset introduces the runtime  chain creation logic into the
orderer.  It's presently only exercised via unit tests, pending the
creation of bdd and sample clients to exercise this behavior.

The chain creation checking is still somewhat limited, and does not
enforce the sorts of keys which can be created, or do real policy
evaluation at this point, but the plug points are there.

Change-Id: I44b904884d81f1e6e279c7214238950b8c6b059f
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>