FAB-1046 Gossip internal identity mapper
The current gossip component assumes the upper layers are able to fetch
each remote peer's certificate by its corresponding PKI-ID.
This is true for 0.5 but this isn't the case for v1,
so this commit makes each peer send the other peer its certificate
and each peer creates a mapping between PKIid-->certificate.

This commit adds an identity mapper that is used for the comm layer
and the gossip layer to map pkiIDs to certificates, in order to
be able to verify messages signed by peers but not having the
certificates inside the messages.

Change-Id: Iab1c99929515c01704ffb0295dca05ccefd8793f
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>