Add interactive asset management demo
This adds an expanded, interactive asset management demo that uses the
preexisting example chaincode "asset_management.go".

Rather than a single application which performs all of the asset
administration, assignments, and transfers, three applications now exist
which perform each of those operations. The third application offers a
command line interface so the user can actually invoke specific asset
transfers of his/her choosing.

A detailed description of these applications and their usage can be found
in the README.

A video demonstration of the applications running can be viewed here:
https://www.youtube.com/watch?v=PCZrFAKXlTE.
Note that this video link is only present in this description and was not
added to any of the committed files.

Change-Id: I4da4bcf99c4829c59bd4659954013c4d42d9930f
Signed-off-by: Mike Zaccardo <mike.zaccardo@cloudsoftcorp.com>