FAB-1046 Gossip identity learning
The current gossip component assumes the upper layers are able to fetch
each remote peer's certificate by its corresponding PKI-ID.
This is true for 0.5 but this isn't the case for v1,
so this commit makes each peer send the other peer its certificate
and each peer creates a mapping between PKIid-->certificate.

The gossip layer now disseminates certificates of peers in the following ways:
1) During the handshake with a remote peer
2) At initialization time, there is a configurable period in which
   the cert is appended to the AliveMessage
3) Peers can now exchange with remote peers certificates by
   comparing the pkiIDs of the certificates. This works with
   the standard built-in pull mechanism.

This commit also moves the comm layer into the gossip, instead
of what was before- where the gossip was given an instance of
the comm.

Change-Id: I7fc3bb0fc59911759566a4b8a9759281ae843f08
Signed-off-by: Yacov Manevich <yacovm@il.ibm.com>