[FAB-1253] Allow attempt to send >1MB blocks to kafka
This is a quick fix to unblock using the Kafka
orderer. Your kafka instance must also be configured
with a corresponding message.max.bytes property in
server.properties, otherwise, the server will reject
the message.

This changeset makes the default 100MiB.

We might want to add this to the orderer config.yml,
but I think it's important to get this quick fix out
into master for downstream users.

Change-Id: Ib4bffbd8afdc229e64d43f011f09d942e0cd5b10
Signed-off-by: Luis Sanchez <sanchezl@us.ibm.com>