[FAB-998] Add new-chain sample client
https://jira.hyperledger.org/browse/FAB-998

This changeset introduces a sample client that sends a single envelope
via the broadcast RPC to the ordering service. This envelope contains a
configuration envelope that calls for the creation of a new chain.

The client is invoked as follows:

$ broadcast_config -cmd new-chain -chainID newChainID

As is the case with the bd_counter sample client, the user can also set
the logging level (-loglevel) and the address of the RPC server
(-server).  The user may also set the backing creation policy
(-creationPolicy).

This changeset also adds chainID support for the deliver_stdout and the
broadcast_timestamp sample clients.

Change-Id: I782eb77c45f5c8d91c75c4440e47a20ab76a560e
Signed-off-by: Kostas Christidis <kostas@christidis.io>
Signed-off-by: Jason Yellick <jyellick@us.ibm.com>