FAB-1256 remove anchor of DefaultChain from peer
https://jira.hyperledger.org/browse/FAB-1256

DefaultChain was the chain that was holding together
the flows through the fabric and was a place holder for
full fledged chains implementation. We now remove that
   . force chainID to be passed through the system, top down
   . use the chainID as the ledgername (note the panic
     in GetLedger)

Pending "join" support, the system still require a single
ledger to be created but that is now created by specifying
the chain at the top level (unit tests, CLI in "peer node start",
etc).

Once "join" is added, this scaffolding can be removed from CLI
(the chaincode unit tests will likely continue to use it) when
the complete chain initialization will be "peer join" configuration
driven

Change-Id: Ie3b64db4b9030a4cb695fb3e1075822b55a129d1
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>