ECDSA Signature malleability resistance
This change-set introduces ECDSA Signature malleability resistance.
ECDSA signatures do not have unique representation and this can facilitate
replay attacks and more. In order to have a unique representation,
this change-set forses BCCSP to generate and accept only signatures
with low-S.
Bitcoin has also addressed this issue with the following BIP:
https://github.com/bitcoin/bips/blob/master/bip-0062.mediawiki
Before merging this change-set, we need to ensure that client-sdks
generates signatures properly in order to avoid massive rejection
of transactions.

This change-set comes in the context of:
https://jira.hyperledger.org/browse/FAB-1276

Change-Id: I83a7a9406ef3551447e1f6540330d3199c0f517e
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>