FAB-1281 add escc, vscc, policy to lccc
https://jira.hyperledger.org/browse/FAB-1281

In addition to name and deployment spec, escc, vscc and policy
needs to be set on the chaincode at deployment time.

This also removes use of the Table structure to basic GetState
PutState to store a chaincode. A go structure is defined for
it and serialized using proto.

Note that "chainid" is not stored. The lccc is likely instantiated
on each chain's ledger and the chain context is clear. The chain
is still passed as params in the functions (ie, no change there)
 in case its needed for chain specific accesses.

Change-Id: Idef7de7381f3a5552dc9acd80c203599a7f9c544
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>