Rework of MSP (config and factories)
This change-set introduces a local MSP and per-chain MSPs for the peer to use.
The local MSP is configured from local config files and is used to get the
signing identity for the peer. Per-chain MSPs are used on each chain to
validate creators/signatures on the various messages. Additionally, in
msp/configfilegenerator there is a sample program file that shows how the
client config may be built out of certificates and keys. It can be removed
later on but it is useful for now to show how to construct that file.

Change-Id: Id9b92a96800a67b27124cc98a6228ad3e03f531a
Signed-off-by: Alessandro Sorniotti <ale.linux@sopit.net>