FAB-1295 - multichannel use chain in chain namespace
https://jira.hyperledger.org/browse/FAB-1295

This changeset uses the passed chainID (previous change set
prepared the way by removing DefaultChain and passing
chain ID top down) as as part of the namespace when registering
and launching chaincodes. With that single stroke, the chaincode
framework is ready for multichain.

This change applies to both user and system chaincodes. The endorser
provides the ledger simulation context for the proposal, outside of
chaincodes, so the chaincodes naturally execute on the chain for the
proposal. Given the chaincode execution uses the the ledger context
supplied to it, all is needed a different container for a chaincode
   . to avoid name collision between chaincodes
   . more secure with no fear of leakage

This changeset deliberately does not excercise multichain capabilities.
The intent is to make sure the changes do not break current code.

The next changeset will provide just test cases for executing chaincodes
on multiple chains.

Change-Id: Icab52e015e8d6373ef4056905695e28d8d58b3e5
Signed-off-by: Srinivasan Muralidharan <muralisr@us.ibm.com>