FAB-1140 Ledger History Database framework
This is an initial check-in that will
1) create the history database in couchDB if it does not exist.
2) add the history commit to the LEDGER Commit if history is enabled
3) stores the history in the history database

The History functionality is not enabled unless both
couchDB and History are enabled in the config.

Note that tests will be added to validate the writes
and the actual data written to the database in future
changes.  The tests to validate history is not possible
until the query APIs are put in place in future changes.

Change-Id: Id207007ab5faae957c1e05234e441566a116ea33
Signed-off-by: Mari Wade <mariwade@us.ibm.com>