Copying shim/crypto to new a package
This change-set is the first step in the process of removing
shim/crypto and the dependencies from the ChaincodeStubInterface.

The change-set introduces a new top-level folder called
accesscontrol that has the same content of shim/crypto
plus a file containing some API.

Subsequent change-sets will remove shim/crypto and ensure
that the chaincode example can still work.

Change-Id: I8b1c701fd708e497dc6d95dbada5619979d05964
Signed-off-by: Angelo De Caro <adc@zurich.ibm.com>