[FAB-1359] Drop custom flags for Kafka orderer
https://jira.hyperledger.org/browse/FAB-1359

As we are slowly moving to a setup where the behavior of the orderer is
controlled by the config options captured in the genesis block and the
orderer YAML file (and their overrides via ENV vars), it's time to drop
the flag support that the Kafka orderer provided.

This changeset then:
1. Removes all flags from the Kafka orderer.
2. Adds a "verbose" option to the YAML file to control logging for the
package that we use to interact with the Kafka cluster (sarama).

Additionally it:
3. Prefixes all test-related variables with "test" so as to make tests
more legible and remove ambiguity.
4. Updates the test config object to match the keys of the actual config
object.
5. Adds a helper test envelope constructor function, and moves some
test-related functions around in an effort to consolidate files.

Change-Id: Id749d0b88f62a4212854e18b8c469d90fe2f6877
Signed-off-by: Kostas Christidis <kostas@christidis.io>